Subnetworks

Functions to pull out sub-models from full models. Intended for use in testing or when there's an interest in a subregion.

Usage

Define a SubNetworkDef struct, then call generate_subnetwork. Once this is done, StreamModel and FlowRegime structs can be created directly from the SubNetworkDef struct

Documentation

NitrateNetworkModel.SubNetworkDef — Type

SubNetworkDef(
    network_file::String,
    baseparams_file::String,
    flow_regime_file::String,
    root_node::Int,
    subnetwork_file::String,
    subnetwork_params_file::String,
    subnetwork_flow_regime_file::Union{Nothing, String}
)

Parametric struct to pass to generate_subnetwork.

The final argument, subnetwork_flow_regime_file can be nothing or a path to a flow regime file. If it points to a file, a corresponding scaled flow regime file will be created, or if it is nothing, this will be skipped.

All other arguments are required.

NitrateNetworkModel.generate_subnetwork — Function

generate_subnetwork(subnetworkdef::SubNetworkDef)

Generate files for running the NitrateNetworkModel on a subnetwork of a larger network model.

NitrateNetworkModel.StreamModel — Method

StreamModel(subnetworkdef::SubNetworkDef)

Construct a StreamModel from a SubNetworkDef

NitrateNetworkModel.FlowRegime — Method

FlowRegime(subnetworkdef::SubNetworkDef)

Construct a FlowRegime from a SubNetworkDef